Renamed processor and module/package to drop '3'
