+++ nifi-extension-bundles/nifi-aws-bundle/nifi-aws-kinesis-nar/src/main/resources/META-INF/LICENSE
